requirements.txt
# for building android
buildozer==1.5.0
Cython==3.0.12
setuptools==82.0.1
python-for-android==2024.1.21
    #appdirs
    #colorama
    #build
        #pyproject_hooks

# android implicitly imported requirements
#

# for the app (formatted as dependency tree)
llama-cpp-python==0.3.16
    #typing-extensions>=4.5.0
    #numpy>=1.20.0
    #MarkupSafe>=2.0
    #diskcache>=5.6.1
        #jinja2>=2.11.3
opencv-python==4.13.0.92 # available as recipe opencv
pytesseract==0.3.13
    #packaging>=21.3
    #pillow>=8.0.0
speechrecognition==3.15.1
    #standard-aifc
        #standard-chunk
        #audioop-lts
pyaudio==0.2.14